fix: delay squadjs startup by 10 seconds
We have to wait until the Squad server is initialized and has a full
RCON connection up. This allows SquadJS to be started in tandem with a
Squad docker image in docker compose without conflict.

This is a bit hacky and a healthcheck would be superior, but there's a
lot of instances to check in the Squad log to even verify if the RCON
connection is good to go.

Another potential way of doing this health check is to attempt to login
into the rcon connection ourselves in a loop and then start SquadJS once
that rcon connection succeeds (after closing the RCON connection we
made).